order tests

+++ database/migrations/2020_10_29_174604_create_promos_table.php
@@ -17,7 +17,7 @@ class CreatePromosTable extends Migration
             $table->id();
             $table->string('name');
             $table->enum('type', ['amount_off', 'percent_off']);
-            $table->integer('value');
+            $table->unsignedBigInteger('value');
             $table->string('code');
             $table->timestamps();
         });
